refactor: 게스트 커맨드 서비스를 분리
