+++ src/main/java/umc/cockple/demo/domain/exercise/controller/ExerciseGuestController.java
@@ -9,7 +9,7 @@ import umc.cockple.demo.domain.exercise.dto.participation.ExerciseCancelDTO;
 import umc.cockple.demo.domain.exercise.dto.guest.ExerciseGuestInviteDTO;
 import umc.cockple.demo.domain.exercise.dto.guest.ExerciseMyGuestListDTO;
 import umc.cockple.demo.domain.exercise.service.query.ExerciseGuestQueryService;
-import umc.cockple.demo.domain.exercise.service.command.ExerciseCommandService;
+import umc.cockple.demo.domain.exercise.service.command.ExerciseGuestCommandService;
 import umc.cockple.demo.global.response.BaseResponse;
 import umc.cockple.demo.global.response.code.status.CommonSuccessCode;
 import umc.cockple.demo.global.security.utils.SecurityUtil;
@@ -19,7 +19,7 @@ import umc.cockple.demo.global.security.utils.SecurityUtil;
 @Validated
 public class ExerciseGuestController implements ExerciseGuestApi {
 
-    private final ExerciseCommandService exerciseCommandService;
+    private final ExerciseGuestCommandService exerciseGuestCommandService;
     private final ExerciseGuestQueryService exerciseGuestQueryService;
 
     @Override
@@ -27,7 +27,7 @@ public class ExerciseGuestController implements ExerciseGuestApi {
             Long exerciseId, ExerciseGuestInviteDTO.Request request) {
         Long inviterId = SecurityUtil.getCurrentMemberId();
 
-        ExerciseGuestInviteDTO.Response response = exerciseCommandService.inviteGuest(
+        ExerciseGuestInviteDTO.Response response = exerciseGuestCommandService.inviteGuest(
                 exerciseId, inviterId, request);
 
         return BaseResponse.of(CommonSuccessCode.CREATED, response);
@@ -38,7 +38,7 @@ public class ExerciseGuestController implements ExerciseGuestApi {
             Long exerciseId, Long guestId) {
         Long memberId = SecurityUtil.getCurrentMemberId();
 
-        ExerciseCancelDTO.Response response = exerciseCommandService.cancelGuestInvitation(
+        ExerciseCancelDTO.Response response = exerciseGuestCommandService.cancelGuestInvitation(
                 exerciseId, guestId, memberId);
 
         return BaseResponse.of(CommonSuccessCode.OK, response);
