--- src/main/java/umc/cockple/demo/domain/exercise/service/command/ExerciseCommandService.java
@@ -1,67 +1,0 @@
-package umc.cockple.demo.domain.exercise.service.command;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import umc.cockple.demo.domain.exercise.domain.Exercise;
-import umc.cockple.demo.domain.exercise.domain.Guest;
-import umc.cockple.demo.domain.exercise.dto.participation.ExerciseCancelDTO;
-import umc.cockple.demo.domain.exercise.dto.guest.ExerciseGuestInviteDTO;
-import umc.cockple.demo.domain.exercise.exception.ExerciseErrorCode;
-import umc.cockple.demo.domain.exercise.exception.ExerciseException;
-import umc.cockple.demo.domain.exercise.repository.ExerciseRepository;
-import umc.cockple.demo.domain.exercise.repository.GuestRepository;
-import umc.cockple.demo.domain.exercise.service.command.internal.ExerciseGuestService;
-import umc.cockple.demo.domain.member.domain.Member;
-import umc.cockple.demo.domain.member.repository.MemberRepository;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-@Slf4j
-public class ExerciseCommandService {
-
-    private final ExerciseGuestService exerciseGuestService;
-
-    private final MemberRepository memberRepository;
-    private final ExerciseRepository exerciseRepository;
-    private final GuestRepository guestRepository;
-
-    public ExerciseGuestInviteDTO.Response inviteGuest(Long exerciseId, Long inviterId, ExerciseGuestInviteDTO.Request request) {
-        log.info("게스트 초대 시작 - exerciseId: {}, inviterId: {}, guestName: {}"
-                , exerciseId, inviterId, request.guestName());
-
-        Exercise exercise = findExerciseOrThrow(exerciseId);
-        Member inviter = findMemberOrThrow(inviterId);
-
-        return exerciseGuestService.inviteGuest(exercise, inviter, request);
-    }
-
-    public ExerciseCancelDTO.Response cancelGuestInvitation(Long exerciseId, Long guestId, Long memberId) {
-        log.info("게스트 초대 취소 시작 - exerciseId: {}, guestId: {}, memberId: {}", exerciseId, guestId, memberId);
-
-        Exercise exercise = findExerciseOrThrow(exerciseId);
-        Member member = findMemberOrThrow(memberId);
-        Guest guest = findGuestOrThrow(guestId);
-
-        return exerciseGuestService.cancelGuestInvitation(exercise, guest, member);
-    }
-
-    // ========== 조회 메서드 ==========
-
-    private Member findMemberOrThrow(Long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.MEMBER_NOT_FOUND));
-    }
-
-    private Exercise findExerciseOrThrow(Long exerciseId) {
-        return exerciseRepository.findById(exerciseId)
-                .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.EXERCISE_NOT_FOUND));
-    }
-
-    private Guest findGuestOrThrow(Long guestId) {
-        return guestRepository.findById(guestId)
-                .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.GUEST_NOT_FOUND));
-    }
-}
